fetch input schema from job input forms, improve loading
