+++ web-app/src/app/app/agents/[agentId]/jobs/@right/components/create-job-section.tsx
@@ -4,20 +4,17 @@ import { useTranslations } from "next-intl";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { getDescription } from "@/lib/db/extension/agent";
 import { AgentWithRelations } from "@/lib/db/types/agent.types";
-import { JobInputsDataSchemaType } from "@/lib/job-input";
 
 import { JobInputsForm } from "./job-input";
 
 interface CreateJobSectionProps {
   agent: AgentWithRelations;
   agentPricing: number;
-  inputSchema: JobInputsDataSchemaType;
 }
 
 export default function CreateJobSection({
   agent,
   agentPricing,
-  inputSchema,
 }: CreateJobSectionProps) {
   const t = useTranslations("App.Agents.Jobs.CreateJob");
 
@@ -47,11 +44,7 @@ export default function CreateJobSection({
             <h1 className="text-xl font-bold">{t("Input.title")}</h1>
             <p className="text-base">{t("Input.description")}</p>
           </div>
-          <JobInputsForm
-            agentId={agent.id}
-            agentPricing={agentPricing}
-            jobInputsDataSchema={inputSchema}
-          />
+          <JobInputsForm agentId={agent.id} agentPricing={agentPricing} />
         </div>
       </ScrollArea>
     </div>
