+++ web-app/src/app/app/agents/[agentId]/jobs/@right/components/job-input/job-inputs-form.tsx
@@ -1,137 +1,62 @@
-"use client";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { Loader2 } from "lucide-react";
-import { usePathname } from "next/navigation";
-import { useTranslations } from "next-intl";
-import { SubmitHandler, useForm } from "react-hook-form";
-import { toast } from "sonner";
+import { Suspense } from "react";
 
-import { Button } from "@/components/ui/button";
-import { Form } from "@/components/ui/form";
-import { startJobWithInputData } from "@/lib/actions/job.actions";
-import {
-  defaultValues,
-  JobInputsDataSchemaType,
-  jobInputsFormSchema,
-  JobInputsFormSchemaType,
-} from "@/lib/job-input";
-import { cn } from "@/lib/utils";
+import { Skeleton } from "@/components/ui/skeleton";
+import { getAgentInputSchema } from "@/lib/db/services/agent.service";
 
-import JobInput from "./job-input";
-import { useRouterPush } from "./util";
+import JobInputsFormClient from "./job-inputs-form.client";
 
 interface JobInputsFormProps {
   agentId: string;
   agentPricing: number;
-  jobInputsDataSchema: JobInputsDataSchemaType;
   className?: string | undefined;
-}
-
-function transformFormData(values: JobInputsFormSchemaType) {
-  return Object.entries(values).reduce(
-    (acc, [key, value]) => {
-      if (value === null) return acc;
-
-      if (
-        Array.isArray(value) &&
-        value.every((item) => typeof item === "string")
-      ) {
-        const numArray = value.map((v) => Number(v)).filter((n) => !isNaN(n));
-        if (numArray.length === value.length) {
-          acc[key] = numArray;
-          return acc;
-        }
-      }
-
-      if (
-        typeof value === "string" ||
-        typeof value === "number" ||
-        typeof value === "boolean"
-      ) {
-        acc[key] = value;
-      }
-
-      return acc;
-    },
-    {} as Record<string, string | number | boolean | number[]>,
-  );
 }
 
 export default function JobInputsForm({
   agentId,
   agentPricing,
-  jobInputsDataSchema,
   className,
 }: JobInputsFormProps) {
-  const { input_data } = jobInputsDataSchema;
-  const t = useTranslations("Library.JobInput.Form");
-  const form = useForm<JobInputsFormSchemaType>({
-    resolver: zodResolver(jobInputsFormSchema(input_data, t)),
-    defaultValues: defaultValues(input_data),
-  });
-  const push = useRouterPush();
-  const pathname = usePathname();
+  return (
+    <Suspense fallback={<JobInputsFormSkeleton />}>
+      <JobInputsFormInner
+        agentId={agentId}
+        agentPricing={agentPricing}
+        className={className}
+      />
+    </Suspense>
+  );
+}
 
-  // Then replace your existing handleSubmit function with this:
-  const handleSubmit: SubmitHandler<JobInputsFormSchemaType> = async (
-    values,
-  ) => {
-    try {
-      // Transform input data to match expected type
-      // Filter out null values and ensure arrays are of correct type
-      const transformedInputData = transformFormData(values);
-      const result = await startJobWithInputData({
-        agentId: agentId,
-        maxAcceptedCreditCost: agentPricing,
-        inputData: transformedInputData,
-      });
-
-      if (result.success && result.data?.jobId) {
-        form.reset();
-        await push(`${pathname}/${result.data.jobId}`);
-      }
-    } catch (error) {
-      console.error(error);
-      toast.error(t("error"));
-    }
-  };
-
-  const handleClear = () => {
-    form.reset();
-  };
+async function JobInputsFormInner({
+  agentId,
+  agentPricing,
+  className,
+}: JobInputsFormProps) {
+  const inputSchema = await getAgentInputSchema(agentId);
 
   return (
-    <Form {...form}>
-      <form onSubmit={form.handleSubmit(handleSubmit)}>
-        <fieldset
-          disabled={form.formState.isSubmitting}
-          className={cn("flex flex-1 flex-col gap-6", className)}
-        >
-          {input_data.map((jobInputSchema) => (
-            <JobInput
-              key={jobInputSchema.id}
-              form={form}
-              jobInputSchema={jobInputSchema}
-            />
-          ))}
-          <div className="flex items-center justify-between gap-2">
-            <Button type="reset" variant="outline" onClick={handleClear}>
-              {t("clear")}
-            </Button>
-            <div className="flex items-center gap-2">
-              <div className="text-muted-foreground text-sm">
-                {t("price", { price: agentPricing })}
-              </div>
-              <Button type="submit" disabled={form.formState.isSubmitting}>
-                {form.formState.isSubmitting && (
-                  <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                )}
-                {t("submit")}
-              </Button>
-            </div>
-          </div>
-        </fieldset>
-      </form>
-    </Form>
+    <JobInputsFormClient
+      agentId={agentId}
+      agentPricing={agentPricing}
+      jobInputsDataSchema={inputSchema}
+      className={className}
+    />
+  );
+}
+
+function JobInputsFormSkeleton() {
+  return (
+    <div className="flex w-full flex-col gap-2">
+      {Array.from({ length: 3 }).map((_, index) => (
+        <div key={index} className="flex flex-col gap-2">
+          <Skeleton className="h-4 w-24" />
+          <Skeleton className="h-8 w-full" />
+        </div>
+      ))}
+      <div className="flex items-center justify-between">
+        <Skeleton className="h-8 w-16" />
+        <Skeleton className="h-8 w-16" />
+      </div>
+    </div>
   );
 }
